requirements.txt
aioice==0.9.0
aiortc==1.5.0
altair==5.1.2
attrs==23.1.0
av==10.0.0
blinker==1.6.3
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
contourpy==1.1.1
cryptography==41.0.4
cycler==0.12.1
dacite==1.8.1
dnspython==2.4.2
filelock==3.12.4
fonttools==4.43.1
fsspec==2023.10.0
gitdb==4.0.10
GitPython==3.1.38
google-crc32c==1.5.0
graphviz==0.20.1
htmlmin==0.1.12
idna==3.4
ifaddr==0.2.0
ImageHash==4.3.1
importlib-metadata==6.8.0
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
llvmlite==0.41.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.3
mdurl==0.1.2
mpmath==1.3.0
multimethod==1.10
networkx==3.1
numba==0.58.0
numpy==1.25.2
packaging==23.2
pandas==2.0.3
pandas-profiling==3.6.6
patsy==0.5.3
phik==0.12.3
Pillow==10.1.0
plotly==5.17.0
plotly-express==0.4.1
protobuf==4.24.4
psycopg2-binary==2.9.9
pyarrow==13.0.0
pycparser==2.21
pydantic==1.10.13
pydeck==0.8.1b0
pyee==11.0.1
Pygments==2.16.1
pylibsrtp==0.8.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rich==13.6.0
rpds-py==0.10.6
scikit-learn==1.3.1
scipy==1.11.3
seaborn==0.12.2
six==1.16.0
smmap==5.0.1
statsmodels==0.14.0
stqdm==0.0.5
streamlit==1.27.2
streamlit-pandas-profiling==0.1.3
streamlit-webrtc==0.47.1
sympy==1.12
tangled-up-in-unicode==0.2.0
tenacity==8.2.3
threadpoolctl==3.2.0
toml==0.10.2
toolz==0.12.0
torch==2.1.0
torch-summary==1.4.5
torchviz==0.0.2
tornado==6.3.3
tqdm==4.66.1
typeguard==4.1.5
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.1
urllib3==2.0.7
validators==0.22.0
visions==0.7.5
watchdog==3.0.0
wordcloud==1.9.2
ydata-profiling==4.6.0
zipp==3.17.0
